Refactor encoding to use uint32 globally instead of uint64
Changed all size/count/offset fields from uint64 to uint32 (~4B limit instead of 16EB).
This is more appropriate for the scale of this LSM database implementation.

Core changes:
- common/encoding.go: Removed WriteUint64/ReadUint64 helpers
- common/types.go: Entry.Seq, keyLen, valueLen now uint32 (header 1+4+4+4=13 bytes, was 25)
- bitmap: numBits and index parameters now uint32
- bloom filter: m (bits) parameter now uint32
- sstable/IndexEntry: BlockOffset and keyLen now uint32
- sstable/Footer: FilterOffset, IndexOffset, EntryCount now uint32 (footer 12 bytes, was 24)
- memtable: sequence numbers now uint32
- db: nextSeq now uint32

Documentation updates:
- Updated all encoding layout diagrams to show uint32 sizes
- Entry Layout: uint8 + 3×uint32 + data
- IndexEntry Layout: 2×uint32 + data
- Footer: reduced from 24 to 12 bytes
- Bloom filter format: 2×uint32 + bitmap data

Note: Tests still need updating to use uint32 - will be fixed in follow-up commit.

+++ internal/bitmap/bitmap_interface.go
@@ -1,15 +1,16 @@
 package bitmap
 
 // Bitmap is a set interface backed by a space-efficient bit array.
+// We will use it to support a bloom filter.
 type Bitmap interface {
 	// Add sets the bit at position i to 1 (adds i to the set).
-	Add(i uint64)
+	Add(i uint32)
 
 	// Remove sets the bit at position i to 0 (removes i from the set).
-	Remove(i uint64)
+	Remove(i uint32)
 
 	// Contains returns true if bit at position i is set (i is in the set).
-	Contains(i uint64) bool
+	Contains(i uint32) bool
 
 	// Bytes returns the underlying byte array.
 	Bytes() []byte
